Added "Add Dteq and Add Load functions" Need to complete datavalidation and calcualate behavior and link up enter key to add buttons

+++ EDTLibrary/TypeTables/VoltageType.cs
@@ -7,8 +7,8 @@ namespace EDTLibrary.TypeTables
     public class VoltageType
     {
         public int Id { get; set; }
-        public int Voltage { get; set; }
-        public int Phase { get; set; }
+        public double Voltage { get; set; }
+        public double Phase { get; set; }
         public string Category { get; set; }
     }
 }
